+++ R/components-camera.R
@@ -20,29 +20,20 @@ setMethod("initialize", "componentsCamera",
 #' @param onlyIsotopes Logical value. If \code{TRUE} only isotopes are
 #'   considered when generating components (faster). Corresponds to \code{quick}
 #'   argument of \code{\link[CAMERA:annotate-methods]{CAMERA::annotate}}.
-#' @param minSize The minimum size of a component. Smaller components than this
-#'   size will be removed. See note below.
-#' @param ubiquitous If \code{TRUE}: ensure that all feature groups within a
-#'   component are present in the same analyses. See note below.
-#'
-#' @note For \code{generateComponentsCAMERA}: the \code{minSize} and
-#'   \code{ubiquitous} arguments provide additional filtering functionality not
-#'   provided by \pkg{CAMERA}. Consequently, their default values may produce
-#'   different results compared to regular output from \code{CAMERA}.
 #'
 #' @references \addCitations{CAMERA}{1}
 #'
 #' @rdname component-generation
 #' @export
 generateComponentsCAMERA <- function(fGroups, ionization, onlyIsotopes = FALSE,
-                                     minSize = 2, ubiquitous = TRUE, extraOpts = NULL)
+                                     minSize = 2, relMinReplicates = 0.5, extraOpts = NULL)
 {
     ac <- checkmate::makeAssertCollection()
     checkmate::assertClass(fGroups, "featureGroups", add = ac)
     checkmate::assertChoice(ionization, c("positive", "negative"), add = ac)
     checkmate::assertFlag(onlyIsotopes, add = ac)
-    checkmate::assertCount(minSize, add = ac)
-    checkmate::assertFlag(ubiquitous, add = ac)
+    checkmate::assertCount(minSize, positive = TRUE, add = ac)
+    checkmate::assertNumber(relMinReplicates, lower = 0, finite = TRUE, add = ac)
     checkmate::assertList(extraOpts, any.missing = FALSE, names = "unique", null.ok = TRUE, add = ac)
     checkmate::reportAssertions(ac)
 
@@ -50,7 +41,7 @@ generateComponentsCAMERA <- function(fGroups, ionization, onlyIsotopes = FALSE,
         return(componentsCamera(componentInfo = data.table(), components = list(),
                                 xsa = new("xsAnnotate")))
 
-    hash <- makeHash(fGroups, ionization, onlyIsotopes, minSize, ubiquitous, extraOpts)
+    hash <- makeHash(fGroups, ionization, onlyIsotopes, minSize, relMinReplicates, extraOpts)
     cd <- loadCacheData("componentsCAMERA", hash)
     if (!is.null(cd))
         return(cd)
@@ -121,18 +112,25 @@ generateComponentsCAMERA <- function(fGroups, ionization, onlyIsotopes = FALSE,
     comps <- pruneList(setNames(sapply(seq_along(comps), function(cmpi)
     {
         cmp <- comps[[cmpi]]
-        if (ubiquitous)
+        if (relMinReplicates > 0)
         {
             fgCmp <- removeEmptyAnalyses(fGroups[, cmp$group])
-            fgCmp <- minAnalysesFilter(fgCmp, relThreshold = 1, verbose = FALSE)
+            fgCmp <- minReplicatesFilter(fgCmp, relThreshold = relMinReplicates, verbose = FALSE)
             cmp <- cmp[group %in% names(fgCmp)]
         }
-        if (minSize > 0 && nrow(cmp) < minSize)
+        if (nrow(cmp) < minSize)
             return(cmp[0])
             
         anai <- an@psSamples[[cmpi]]
         cmp[, "intensity" := unlist(gTable[anai, cmp$group, with = FALSE])]
     }, simplify = FALSE), cnames), checkZeroRows = TRUE)
+    
+    if (length(comps) != length(cmpanalyses))
+    {
+        # component(s) got filtered out, update
+        cmpanalyses <- cmpanalyses[names(comps)] # update
+        names(comps) <- paste0("CMP", seq_along(comps))
+    }
     
     rets <- lapply(comps, function(cm) gInfo[cm$group, "rts"])
     if (!is.null(adTable))
@@ -150,10 +148,8 @@ generateComponentsCAMERA <- function(fGroups, ionization, onlyIsotopes = FALSE,
 
     cInfo <- data.table(name = names(comps), cmp_ret = sapply(rets, mean),
                         cmp_retsd = sapply(rets, sd), neutral_mass = Ms,
-                        analysis = cmpanalyses[names(comps)], size = sapply(comps, nrow))
+                        analysis = cmpanalyses, size = sapply(comps, nrow))
 
-
-    # UNDONE: keep n=1 sized components?
 
     ret <- componentsCamera(xsa = an, components = comps, componentInfo = cInfo)
     saveCacheData("componentsCAMERA", ret, hash)