+++ tests/testthat/test-components.R
@@ -1,15 +1,20 @@
 context("components")
 
+# take a blank and standard to have two different replicate groups
 # set localMZRange=0 to keep isotopes
-fGroups <- getTestFGroups(getTestAnaInfo()[4:5, ], localMZRange = 0)
+fGroups <- getTestFGroups(getTestAnaInfo()[3:4, ], localMZRange = 0)
 # reduced set for CAMERA/RAMClustR; for nontarget we keep all to get less common homologues
 fGroupsSimple <- fGroups[, 1:50]
 
 
 # fix seed for reproducible clustering, suppress warnings about <5 samples
 withr::with_seed(20, suppressWarnings(compsRC <- generateComponents(fGroupsSimple, "ramclustr", ionization = "positive")))
+withr::with_seed(20, suppressWarnings(compsRCMR <- generateComponents(fGroupsSimple, "ramclustr",
+                                                                      ionization = "positive", relMinReplicates = 1)))
 # UNDONE: getting unknown NaN warnings here...
 suppressWarnings(compsCAM <- generateComponents(fGroupsSimple, "camera", ionization = "positive"))
+suppressWarnings(compsCAMMR <- generateComponents(fGroupsSimple, "camera", ionization = "positive", relMinReplicates = 1))
+suppressWarnings(compsCAMSize <- generateComponents(fGroupsSimple, "camera", ionization = "positive", minSize = 3))
 compsNT <- generateComponents(fGroups, "nontarget", ionization = "positive")
 compsInt <- generateComponents(fGroupsSimple, "intclust", average = FALSE) # no averaging: only one rep group
 fGroupsEmpty <- getEmptyTestFGroups()
@@ -20,10 +25,16 @@ test_that("components generation works", {
     expect_known_value(list(componentTable(compsRC), componentInfo(compsRC)), testFile("components-rc"))
     expect_known_value(list(componentTable(compsCAM), componentInfo(compsCAM)), testFile("components-cam"))
     expect_known_value(compsInt, testFile("components-int"))
+    
     expect_length(compsEmpty, 0)
     expect_length(generateComponents(fGroupsEmpty, "ramclustr", ionization = "positive"), 0)
     expect_length(generateComponents(fGroupsEmpty, "camera", ionization = "positive"), 0)
     expect_length(generateComponents(fGroupsEmpty, "intclust"), 0)
+
+    expect_lt(length(compsRCMR), length(compsRC))
+    expect_lt(length(compsCAMMR), length(compsCAM))
+    expect_gte(min(componentInfo(compsCAMSize)$size), 3)
+    
     skip_if(length(compsNT) == 0)
     expect_known_value(compsNT, testFile("components-nt"))
     expect_length(generateComponents(fGroupsEmpty, "nontarget", ionization = "positive"), 0)